feat: improve the create, change, terminate dod settlement smart contract process
